Move range and sequence tests into the corresponding modules
